extract abstractions
