+++ lib/layers/data/data_sources/abstractions/local_data_source.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/material.dart';
+
+abstract class LocalDataSource {
+  ThemeMode getThemeMode();
+
+  Future<void> setThemeMode(ThemeMode themeMode);
+
+  Locale getLocale();
+
+  Future<void> setLocale(Locale locale);
+}
