+++ lib/layers/data/data_sources/abstractions/remote_user_data_source.dart
@@ -1,0 +1,5 @@
+import 'package:supermedia/layers/data/models/user_model.dart';
+
+abstract class RemoteUserDataSource {
+  Future<UserModel> signUp(String email, String password);
+}
