+++ lib/layers/data/data_sources/local/local_data_source.dart
@@ -1,16 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:supermedia/di/app_module.dart';
+import 'package:supermedia/layers/data/data_sources/abstractions/local_data_source.dart';
 
-abstract class LocalDataSource {
-  ThemeMode getThemeMode();
-
-  Future<void> setThemeMode(ThemeMode themeMode);
-
-  Locale getLocale();
-
-  Future<void> setLocale(Locale locale);
-}
 
 class LocalDataSourceImpl implements LocalDataSource {
   static const String _themeModeKey = 'theme_mode';
