+++ lib/layers/data/data_sources/remote/remote_user_data_source.dart
@@ -1,12 +1,10 @@
 import 'package:supermedia/common/exceptions/signup_exception.dart';
 import 'package:supermedia/common/utils/app_localization.dart';
 import 'package:supermedia/di/app_module.dart';
+import 'package:supermedia/layers/data/data_sources/abstractions/remote_user_data_source.dart';
 import 'package:supermedia/layers/data/http_client/http_client.dart';
 import 'package:supermedia/layers/data/models/user_model.dart';
 
-abstract class RemoteUserDataSource {
-  Future<UserModel> signUp(String email, String password);
-}
 
 class RemoteUserDataSourceImpl implements RemoteUserDataSource {
   final IHttpClient _httpClient = locator<IHttpClient>();
