+++ lib/layers/domain/use_cases/abstractoins/signup_use_case.dart
@@ -1,0 +1,5 @@
+import 'package:supermedia/layers/domain/entities/user.dart';
+
+abstract class SignUpUseCase {
+  Future<User> signUp(String email, String password);
+}
