+++ lib/layers/domain/use_cases/signup_use_case.dart
@@ -3,10 +3,7 @@ import 'package:supermedia/common/utils/app_localization.dart';
 import 'package:supermedia/di/app_module.dart';
 import 'package:supermedia/layers/domain/entities/user.dart';
 import 'package:supermedia/layers/domain/repositories/user_repository.dart';
-
-abstract class SignUpUseCase {
-  Future<User> signUp(String email, String password);
-}
+import 'package:supermedia/layers/domain/use_cases/abstractoins/signup_use_case.dart';
 
 class SignUpUseCaseImpl implements SignUpUseCase {
   final UserRepository _userRepository = locator<UserRepository>();
