+++ lib/layers/presentation/auth/signup/bloc/signup_bloc.dart
@@ -2,7 +2,7 @@ import 'package:equatable/equatable.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:supermedia/di/app_module.dart';
 import 'package:supermedia/layers/domain/entities/user.dart';
-import 'package:supermedia/layers/domain/use_cases/signup_use_case.dart';
+import 'package:supermedia/layers/domain/use_cases/abstractoins/signup_use_case.dart';
 
 part 'signup_event.dart';
 part 'signup_state.dart';
