+++ lib/main.dart
@@ -7,7 +7,6 @@ import 'package:supermedia/common/theme/app_theme_bloc.dart';
 import 'package:supermedia/common/utils/app_localization.dart';
 import 'package:supermedia/di/app_module.dart';
 import 'package:supermedia/layers/presentation/auth/signup/screens/signup_screen.dart';
-import 'package:supermedia/layers/presentation/settings/screens/settings_screen.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
